feat: create ExamTypePageViewModel
